[Add] #241 - UserInfo DTO 생성
